[FIX]: cycle issue
