--- src/store/redux/combine-reducers.ts
@@ -1,8 +1,0 @@
-import {combineReducers} from '@reduxjs/toolkit';
-import {demoReducer} from './slices';
-
-
-//! Combine all reducers and export
-export const allCombineReducers = combineReducers({
-  demoReducer,
-});
