+++ src/store/redux/index.tsx
@@ -2,11 +2,12 @@ import React from 'react';
 import {Text} from 'react-native';
 import {Provider as ReduxStoreProvider} from 'react-redux';
 import {PersistGate} from 'redux-persist/integration/react';
-import {persistor, store} from './utils';
+import {persistor, reduxStore} from './redux-store';
+
 
 const PersistedReduxProvider = ({ children }: React.PropsWithChildren) => {
   return (
-    <ReduxStoreProvider store={store}>
+    <ReduxStoreProvider store={reduxStore}>
       <PersistGate loading={<Text>Loading...</Text>} persistor={persistor}>
         {children}
       </PersistGate>
