--- src/store/redux/slices/index.ts
@@ -1,1 +1,0 @@
-export * from './demo.slice';
